+++ reproc/src/options.c
@@ -2,30 +2,53 @@
 
 #include "error.h"
 
+static bool redirect_is_set(reproc_redirect redirect)
+{
+  return redirect.type || redirect.handle || redirect.file || redirect.path;
+}
+
 static int parse_redirect(reproc_redirect *redirect,
                           REPROC_STREAM stream,
                           bool parent,
                           bool discard,
-                          FILE *file)
+                          FILE *file,
+                          const char *path)
 {
+  ASSERT(redirect);
+
   if (file) {
-    ASSERT_EINVAL(!redirect->type && !redirect->handle && !redirect->file);
+    ASSERT_EINVAL(!redirect_is_set(*redirect));
+    ASSERT_EINVAL(!parent && !discard && !path);
     redirect->type = REPROC_REDIRECT_FILE;
     redirect->file = file;
+  }
+
+  if (path) {
+    ASSERT_EINVAL(!redirect_is_set(*redirect));
+    ASSERT_EINVAL(!parent && !discard && !file);
+    redirect->type = REPROC_REDIRECT_PATH;
+    redirect->path = path;
   }
 
   if (redirect->type == REPROC_REDIRECT_HANDLE || redirect->handle) {
     ASSERT_EINVAL(!redirect->type || redirect->type == REPROC_REDIRECT_HANDLE);
     ASSERT_EINVAL(redirect->handle);
-    ASSERT_EINVAL(!redirect->file);
+    ASSERT_EINVAL(!redirect->file && !redirect->path);
     redirect->type = REPROC_REDIRECT_HANDLE;
   }
 
   if (redirect->type == REPROC_REDIRECT_FILE || redirect->file) {
     ASSERT_EINVAL(!redirect->type || redirect->type == REPROC_REDIRECT_FILE);
     ASSERT_EINVAL(redirect->file);
-    ASSERT_EINVAL(!redirect->handle);
+    ASSERT_EINVAL(!redirect->handle && !redirect->path);
     redirect->type = REPROC_REDIRECT_FILE;
+  }
+
+  if (redirect->type == REPROC_REDIRECT_PATH || redirect->path) {
+    ASSERT_EINVAL(!redirect->type || redirect->type == REPROC_REDIRECT_PATH);
+    ASSERT_EINVAL(redirect->path);
+    ASSERT_EINVAL(!redirect->handle && !redirect->file);
+    redirect->type = REPROC_REDIRECT_PATH;
   }
 
   if (!redirect->type) {
@@ -51,21 +74,22 @@ int parse_options(reproc_options *options, const char *const *argv)
   int r = -1;
 
   r = parse_redirect(&options->redirect.in, REPROC_STREAM_IN,
-                     options->redirect.parent, options->redirect.discard, NULL);
+                     options->redirect.parent, options->redirect.discard, NULL,
+                     NULL);
   if (r < 0) {
     return r;
   }
 
   r = parse_redirect(&options->redirect.out, REPROC_STREAM_OUT,
                      options->redirect.parent, options->redirect.discard,
-                     options->redirect.file);
+                     options->redirect.file, options->redirect.path);
   if (r < 0) {
     return r;
   }
 
   r = parse_redirect(&options->redirect.err, REPROC_STREAM_ERR,
                      options->redirect.parent, options->redirect.discard,
-                     options->redirect.file);
+                     options->redirect.file, options->redirect.path);
   if (r < 0) {
     return r;
   }